% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tpr_destinations.R
\docType{data}
\name{tpr_destinations_data}
\alias{tpr_destinations_data}
\title{Themepark Destinations}
\format{
## `tpr_destinations_data`
A data frame with 74 rows and 5 columns
}
\source{
<https://api.themeparks.wiki/docs/v1/#/destinations/getDestinations>
}
\usage{
tpr_destinations_data
}
\description{
A snapshot of content results from `tpr_destinations()`
}
\keyword{datasets}
