% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tpr_entity_children.R
\docType{data}
\name{tpr_entity_children_data}
\alias{tpr_entity_children_data}
\title{Themepark Entity Children}
\format{
## `tpr_entity_children_data`
A data frame with 4997 rows and 6 columns
}
\source{
<https://api.themeparks.wiki/docs/v1/#/entities/getEntityChildren>
}
\usage{
tpr_entity_children_data
}
\description{
A snapshot of content results from `tpr_entity_children()`
}
\keyword{datasets}
