% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tpr_entity.R
\docType{data}
\name{tpr_entity_data}
\alias{tpr_entity_data}
\title{Themepark Entity}
\format{
## `tpr_entity_data`
A data frame with 60 rows and 7 columns
}
\source{
<https://api.themeparks.wiki/docs/v1/#/entities/getEntity>
}
\usage{
tpr_entity_data
}
\description{
A snapshot of content results from `tpr_entity()`
}
\keyword{datasets}
