% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tpr_entity_live.R
\docType{data}
\name{tpr_entity_live_data}
\alias{tpr_entity_live_data}
\title{Themepark Entity Live}
\format{
## `tpr_entity_live_data`
A data frame with 3180 rows and 12 columns
}
\source{
<https://api.themeparks.wiki/docs/v1/#/entities/getEntityLiveData>
}
\usage{
tpr_entity_live_data
}
\description{
A snapshot of content results from `tpr_entity_live()`
}
\keyword{datasets}
