% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_destinations.R
\name{tpr_destinations}
\alias{tpr_destinations}
\title{Get destinations}
\usage{
tpr_destinations()
}
\value{
a tibble
}
\description{
Get a list of supported destinations available on 'https://api.themeparks.wiki/'.
}
\details{
This provides both the response and parsed results from the `/destinations` path.
}
\examples{
tpr_destinations()



}
