% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_entity.R
\name{tpr_entity}
\alias{tpr_entity}
\title{Get entity details}
\usage{
tpr_entity(id)
}
\arguments{
\item{id}{GUID or slug string for the entity of interest}
}
\value{
a tibble
}
\description{
Get the full data document for a given entity from 'https://api.themeparks.wiki/'.
    You can supply either a GUID or slug string.
}
\details{
This provides both the response and parsed results from the `/entity/{entityID}` path.
}
\examples{
park_dest = tpr_destinations()$id[[1]]
tpr_entity(park_dest)



}
