% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_entity.R
\name{tpr_entity}
\alias{tpr_entity}
\title{Get entity details}
\usage{
tpr_entity(park)
}
\arguments{
\item{park}{GUID or slug string for the entity of interest from `tpr_destinations()`}
}
\value{
an object of class `themeparks_api`
}
\description{
Get the full data document for a given entity from 'https://api.themeparks.wiki/'.
    You can supply either a GUID or slug string.
}
\details{
This provides both the response and parsed results from the `/entity/{entityID}` path.
}
\examples{
park_dest = tpr_destinations()
tpr_entity(park_dest$content$id[[1]])



}
