% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_entity_children.R
\name{tpr_entity_children}
\alias{tpr_entity_children}
\title{Get entity children details}
\usage{
tpr_entity_children(park)
}
\arguments{
\item{park}{GUID or slug string for the entity of interest}
}
\value{
an object of class `themeparks_api`
}
\description{
Get a list of all the children that belong an entity on
    'https://api.themeparks.wiki/'.
}
\details{
This is recursive, so a destination will
   return all parks and all rides within those parks.
}
\examples{
purrr::map_dfr(tpr_destinations()$content$id[[1]],
    ~tpr_entity_children(.x) \%>\% purrr::pluck("content"))



}
