% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_entity_live.R
\name{tpr_entity_live}
\alias{tpr_entity_live}
\title{Get live entity details}
\usage{
tpr_entity_live(id)
}
\arguments{
\item{id}{GUID or slug string for the entity of interest}
}
\value{
an object of class `themeparks_api`
}
\description{
Get live data (queue times, parade times, etc.) as well as all
    child entities on 'https://api.themeparks.wiki/'.
}
\details{
This is recursive, so a destination will
   return detais for all parks and all rides within those parks.
}
\examples{
purrr::map_dfr(tpr_destinations()$content$id[[1]],
    ~tpr_entity_live(.x) \%>\% purrr::pluck("content"))



}
